adding loader for json-ld
