+++ schemaorg/main/parse/__init__.py
@@ -18,3 +18,4 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 '''
 
 from .recipe import RecipeParser
+
